Extract preprocessing from detector
